github-workflow: bitbake: Use different cache for each machine arch
As of now, GitHub's caching functionality is immutable and cannot be
updated. This means that the first job gets to created the cache
for a specific machine architecture. All following jobs with the same
architecture can make good use of that cache, but all other jobs
must download and compile some or all software.

By introducing caches based on machine architecture, this limitation
won't hit that hard and the jobs can make proper use of the cache.

See https://github.com/actions/cache/blob/main/tips-and-workarounds.md#update-a-cache