recipes-core: Include sources into build
The GPL-2.0-with-classpath-exception license requires to also distribute the source code.
The recipes of the JREs has been adapted in this way.
The sources are getting downloaded and placed into your deploy folder,
which might be situated under build/tmp-glibc/deploy.
Please take a look in the documentation [1] to figure out how to
meet the requirement in your specific build setup.

[1] https://docs.yoctoproject.org/dev-manual/licenses.html

Signed-off-by: Sebastian Veit <sebastian.veit@lucimber.com>