openjdk-8-jdk: alternative fix for duplicate libjli.so
Use the PRIVATE_LIBS variable to tell the packaging code
to ignore libjli.so during its shared library analysis,
so it doesn't flag the duplicate copies. Since the two
copies are not identical, and each copy is private to
the pre-built programs in the package, this should be
safer and cleaner than removing one of the copies.

Signed-off-by: Matt Madison <matt@madison.systems>