libatomic: Remove recipe
The gcc-runtime recipe in openembedded-core provides the
libatomic subpackage.

meta-openjdk-temurin currently supports Yocto 4.3 Nanbield or
newer. Yocto 4.3 uses GCC 13.2.

On the other hand, the libatomic recipe in meta-openjdk-temurin
wants to use the GCC 10.x branch.

As a result, the higher version of the two libatomic1 package
builds would get installed, so the libatomic recipe is not
actually needed.

The recipe as is cannot build now, since:
* the SRC_URI in the libatomic recipe here is not valid now, and
* the oe-core changes made the recipe fail to build in Yocto 5.3

Bottom line: remove the libatomic recipe.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>